requirements.txt
# Core LangChain Dependencies
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.2
langchain-experimental==0.0.49
langgraph==0.0.20

# LLM Providers
openai==1.10.0
anthropic==0.8.1

# Tools and Utilities
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
wikipedia==1.4.0
python-dotenv==1.0.0

# Web scraping and APIs
beautifulsoup4==4.12.2
httpx==0.25.2

# Development and Testing
jupyter==1.0.0
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# Visualization
matplotlib==3.7.2
plotly==5.17.0

# Additional tools
yfinance==0.2.18
tavily-python==0.3.0
